requirements.txt
openai
transformers
torch
langdetect
flask
requests
nltk
scikit-learn
gradio
google-generativeai
speechrecognition
deep-translator